fs_discord_bot/requirements.txt
discord.py
aiohttp
SimConnect
PyInstaller
pytz
